refactor : utils 파일 내 메소드의 위치를 보다 domain-specific 한 방식으로 수정
