+++ src/main/java/webserver/utils/CommonUtils.java
@@ -1,10 +1,13 @@
 package webserver.utils;
 
+import com.google.common.base.Strings;
+import com.google.common.collect.Maps;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class CommonUtils {
 
@@ -36,5 +39,48 @@ public class CommonUtils {
         }
         return result;
     }
+
+    public static Map<String, String> parseValues(String values, String separator) {
+        if (Strings.isNullOrEmpty(values)) {
+            return Maps.newHashMap();
+        }
+
+        String[] tokens = values.split(separator);
+
+        return Arrays.stream(tokens).map(t -> getKeyValue(t, "=")).filter(p -> p != null)
+                .collect(Collectors.toMap(p -> p.getKey(), p -> p.getValue()));
+    }
+
+    private static Pair getKeyValue(String keyValue, String regex) {
+        if (Strings.isNullOrEmpty(keyValue)) {
+            return null;
+        }
+
+        String[] tokens = keyValue.split(regex);
+        if (tokens.length != 2) {
+            return null;
+        }
+
+        return new Pair(tokens[0], tokens[1]);
+    }
+
+    public static class Pair {
+        String key;
+        String value;
+
+        Pair(String key, String value) {
+            this.key = key.trim();
+            this.value = value.trim();
+        }
+
+        public String getKey() {
+            return key;
+        }
+
+        public String getValue() {
+            return value;
+        }
+    }
+
 
 }
